Lots of changes made. Keep Reading for details...
I noticed when using option -D for the config_diff I no longer needed to set a var and debug to print the output
so I've removed it.

Managed to update cisco_create_loop.yml to create and/or delete loopbacks.  There's different modules for creating
the interfaces, adding descriptions, int status and then the L3 details.

I'm also using this as a first attempt at doing github properly, committing changes, including a readme
and meaningful comments.  Also learned the difference between a commit and a push.  They're not the same.